Fix to support GDBM >= 1.14
In this and more recent versions, `gdbm_errno` has been turned into a
macro that is shorthand for `(*gdbm_errno_location ())`.

This patch first tries to attach `gdbm_errno` (for backwards
compatibility), and if that fails it attaches `gdbm_errno_location()`
instead, and defines a function to do the pointer dereferencing.